# shared/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AICloudBuilder API",
    "version": "2.0.0",
    "description": "Spec-Driven API for AICloudBuilder — Terraform, Kubernetes, Docker, ArchMe, and more."
  },
  "servers": [
    {
      "url": "http://localhost:9005",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
...
      "UserPublic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "createdAt",
          "updatedAt"
        ]
      },
      "SignupResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserPublic"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "user",
          "token"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          },
          "details": {
            "nullable": true
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "SignupBody": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 30
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserPublic"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "user",
          "token"
        ]
      },
      "LoginBody": {
        "type": "object",
        "properties": {
          "usernameOrEmail": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          },
          "rememberMe": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "string",
                "enum": [
                  "true"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "false"
                ]
              },
              {
                "nullable": true
              }
            ]
          }
        },
        "required": [
          "usernameOrEmail",
          "password"
        ]
      },
...
      "MeResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserPublic"
          }
        },
        "required": [
          "user"
        ]
      },
...
      "HistoryResponse": {
        "type": "object",
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "nullable": true
            }
          },
          "activities": {
            "type": "array",
            "items": {
              "nullable": true
            }
          },
          "totalSessions": {
            "type": "number"
          },
          "totalActivities": {
            "type": "number"
          }
        },
        "required": [
          "sessions",
          "activities",
          "totalSessions",
          "totalActivities"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/auth/signup": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user/history": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "module",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0,
              "default": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "History",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}